requirements.txt
streamlit
pandas
openpyxl
pillow
streamlit-autorefresh
pdfplumber
html5lib
lxml